Changed all value to double to minimise the lose. Added calculation of cost with sigmoid. Added the method for survivance calculus

+++ src/main/java/alisgroup/titanicmanipulator/Matrix.java
@@ -5,38 +5,38 @@ import java.lang.reflect.Method;
 import java.util.List;
 
 public class Matrix {
-    float matrixX[][];
-    float matrixY[];
+    double matrixX[][];
+    double matrixY[];
 
     public Matrix (){}
-    public Matrix(float[][] matrixX, float[] matrixY) {
+    public Matrix(double[][] matrixX, double[] matrixY) {
         this.matrixX = matrixX;
         this.matrixY = matrixY;
     }
 
     public <Person> Matrix createMatrix(List<Person> persons, Method... methods) throws InvocationTargetException, IllegalAccessException {
-        float[][] X = new float[persons.size()][methods.length];
-        float[] Y = new float[persons.size()];
+        double[][] X = new double[persons.size()][methods.length];
+        double[] Y = new double[persons.size()];
         for(int i = 0; i < persons.size(); i++){
             Person person = persons.get(i);
             for (int j = 0; j < methods.length; j++) {
                 if(j == 0) {X[i][j] = 1;}
-                else {X[i][j] = (float) methods[j-1].invoke(person);}
+                else {X[i][j] = (double) methods[j-1].invoke(person);}
             }
         }
         for(int i = 0; i < persons.size(); i++){
             Person person = persons.get(i);
-            Y[i] = (float) methods[methods.length-1].invoke(person);
+            Y[i] = (double) methods[methods.length-1].invoke(person);
         }
 
         return new Matrix(X, Y);
     }
 
-    public float[][] getMatrixX() {
+    public double[][] getMatrixX() {
         return matrixX;
     }
 
-    public float[] getMatrixY() {
+    public double[] getMatrixY() {
         return matrixY;
     }
 }
